Simplify setup.py
