+++ build.sh
@@ -1,4 +1,3 @@
 #!/bin/bash
-#source ~/.venv/bin/activate
 MAX_JOBS=16 python3 inference/setup.py develop --install-dir=build
